Update backend dependencies and add io module
This commit updates the backend dependencies by adding the ballerina/io module to the Dependencies.toml file. This module is required for the new functionality implemented in the main.bal file.

The main.bal file has been modified to import the ballerina/io module. This module is used in the new resource function `getProductbyID` to print the `productIdParam` value.

These changes are necessary to support the new functionality related to retrieving product details by ID.